requirements_mobile.txt
flet>=0.22.0
llama-cpp-python>=0.2.77
numpy>=1.26.4
Flask>=3.0.3
